+++ genie-web/src/main/java/com/netflix/genie/web/rpc/grpc/interceptors/package-info.java
@@ -22,4 +22,4 @@
  * @author mprimi
  * @since 4.0.0
  */
-package com.netflix.genie.web.rpc.interceptors;
+package com.netflix.genie.web.rpc.grpc.interceptors;
